+++ MdeModulePkg/Bus/Pci/PciBusDxe/PciEnumerator.c
@@ -983,7 +983,7 @@ PciHostBridgeAdjustAllocation (
       // Have no way to get ReqRes, AllocRes & Bar here
       //
       ZeroMem (&AllocFailExtendedData, sizeof (AllocFailExtendedData));
-      AllocFailExtendedData.DevicePathSize = sizeof (EFI_DEVICE_PATH_PROTOCOL);
+      AllocFailExtendedData.DevicePathSize = (UINT16) sizeof (EFI_DEVICE_PATH_PROTOCOL);
       AllocFailExtendedData.DevicePath     = (UINT8 *) PciResNode->PciDev->DevicePath;
       AllocFailExtendedData.Bar            = PciResNode->Bar;
 
@@ -1108,7 +1108,7 @@ ConstructAcpiResourceRequestor (
     //
     if ((Aperture & 0x01) != 0) {
       Ptr->Desc     = ACPI_ADDRESS_SPACE_DESCRIPTOR;
-      Ptr->Len      = sizeof (EFI_ACPI_ADDRESS_SPACE_DESCRIPTOR) - 3;
+      Ptr->Len      = (UINT16) (sizeof (EFI_ACPI_ADDRESS_SPACE_DESCRIPTOR) - 3);
       //
       // Io
       //
